requirements.txt
atproto
google-generativeai
python-dotenv
streamlit
